+++ util-stats/src/test/scala/com/twitter/finagle/stats/DefaultStatsReceiverTest.scala
@@ -3,12 +3,12 @@ package com.twitter.finagle.stats
 import org.scalatest.funsuite.AnyFunSuite
 
 class DefaultStatsReceiverTest extends AnyFunSuite {
-  test("DefaultStatsReceiver has app in dimensional scope and label") {
+  test("DefaultStatsReceiver initializes dimensional scope and label") {
     LoadedStatsReceiver.self = new InMemoryStatsReceiver
     val fooCounter = DefaultStatsReceiver.counter("foo")
     val identity = fooCounter.metadata.toMetricBuilder.get.identity
 
-    assert(identity.dimensionalName == Seq("app", "foo"))
+    assert(identity.dimensionalName == Seq("foo"))
     assert(identity.hierarchicalName == Seq("foo"))
     assert(identity.labels == Map("implementation" -> "app"))
   }